+++ apps/extension/cypress/stubs/tabAssignSelect.tsx
@@ -1,2 +1,0 @@
-const tabAssignSelect = () => null;
-export default tabAssignSelect;